% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_path.R
\name{available_configs}
\alias{available_configs}
\title{List available config identifiers}
\usage{
available_configs()
}
\value{
Character vector of config ids (basenames with `.json.gz`
  stripped).
}
\description{
List available config identifiers
}
