% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_path.R
\name{load_ct}
\alias{load_ct}
\title{Load bundled CDISC Controlled Terminology}
\usage{
load_ct(package = c("sdtm", "adam"))
}
\arguments{
\item{package}{One of `"sdtm"` or `"adam"`.}
}
\value{
Data frame of CT terms.
}
\description{
Returns the decompressed CT terms for one submission standard.
`herald` merges these on top of any sponsor-supplied CT via
`register_ct()`.
}
